JSON encode strings with SetEscapeHTML(false)
When encoding object field names and string literals, create a
json.Encoder with SetEscapeHTML(false) to prevent escaping `&` as
`\u0026` and other ugly work-arounds meant to avoid problems when
embedding JSON in HTML (which we are not doing).  See
https://golang.org/pkg/encoding/json/#Encoder.SetEscapeHTML for more
details.

Those that do want the escaping to happen can set Formatter.EscapeHTML
to true.

Fixes #1